Fix compile break with MSVC 14's standard library
tr1::tuple can be disabled in Visual C++ 2015's standard library by
defining _HAS_TR1_NAMESPACE to 0, which is done by default in C++17
mode. Also, this is not a property of the compiler but the standard
library, and MSVC's standard library is also used by clang on Windows.
Check _CPPLIB_VER instead of _MSC_VER to fix the compile breaks with
MSVC 14.0 as well as clang.